#25238 #25427 #25569 #25609 #25652 Bug fixes.
#25238 Fixes a bug causing duplicates to show up in the type list. 
Previously, the type list showed a raw dump of the shotgun data. This
meant that if you had two types in this list both named "maya anim", you
would have two "maya anim" entries showing up, which is confusing. This
is common if you still use the tank types (or in some cases have
converted from from the tank types). This change collapses type entries
with the same name down to a single entry.

#25569 unix timestamps are now converted before being passed to hook. 
The unix timestamp handling is to mitigate various shortcomings in the
QT/PySide/SG API serialization capabilities and means that all time
stamps in the model data are cached as unix timestamps (number of
seconds since 1 Jan 1970 in UTC time zone -
http://en.wikipedia.org/wiki/Unix_time). This means that all data
currently being returned by the sg model comes back with this time data
format rather than the standard SG Time stamp. (there is no easy way to
slap an interface around the return data). 

This fix converts these unix timestamps back to sg timestamps prior to
passing shotgun data to the action hooks, making it more familiar for
TDs and developers.

#25652 Fixed pixmap:scaled errors.

#25609 Version history now constrains by task and project. 
When displaying the verison history based on an existing publish,
the associated fields to display are now determined based on more
commonality than before - previously, it was only name, type and
associated entity. With this change, project and task are also included.
This means that you can have two publishes with the same type and same
name for the same shot, but with different tasks assocatied, and the
loader will treat their version histories as completely independent.

#25427 Added support for per-profile publish filters. Adds a new
parameter to the config such that you can specify a publish filter for
each tab on the left hand side.

Closes #4