#25877 Fixed issue which can obscure publishes in the main overview.
Fixed a bug which would cause publishes to not show up in the main
overview if there were two publish "streams" with the same name and type
but with different tasks. These would be shown as single entry rather
than two independent streams. This change generates two distinct streams
and also attaches the task name next to the publish name in the display
so that the two different streams are distinguishable from each other.

Closes #5.