#27793 Adds list view mode to loader main UI + brings delegates local.
Introduces a new list mode in the loader which is a more compact
representation of publishes. The delegate widgets have also been moved
into the loader so that it no longer depends on the standard widgets in
the qtwidgets framework. This is a tweak to make it easier for the
loader to switch to the v3 qtwidgets framework when that is released as
part of the file manager changes.

Closes #12