#25280 Fixes bug where clicks on tabs would be ignored
This fixes a bug which meant that the very first click on one of the tabs in the left hand side tree view would sometimes not work correctly. If you for example start the loader and click on a tab, the loader would switch to the tab, but it would be completely blank with no data loaded. 

Closes #29