SMOK-29361 | tk-flame_action improved (#52)
* SMOK-29361 | tk-flame_action imporved

DESCRIPTION

This adds functionnality to the tk-flame part of the Shotgun Loader

Now a user can Create a batch group using the supported clips linked to
a Shot entity. Once this Batch group is created, if the user have the
right context (Classic integration for now with the tk-flame-export),
a Render will trigger a publish export to Shotgun.

There's also more configuration options. Theses options allow the user
to change certain behaviour of the loader without having to play in the
Python code.

Finally, there's more documentation and more debug logging to be able to
do better troubleshooting.

* Regex fix and comment typo

* Add media path root to the configuration

* After CR

* bug fix and review