#43996 Play in sg panel now navigates to media player (#53)
This changes the behaviour of the sg panel so that if you click the "play" icon for a version, it will navigate you to the sg media center page rather than screening room which was previously the case. Back when we originally wrote this code, the media center was not yet released and they did not have proper urls that you can navigate to. This change aligns the behaviour in the panel with what you get in other parts of Shotgun.